+++ src/main/kotlin/id/walt/vclib/Defaults.kt
@@ -3,7 +3,6 @@ package id.walt.vclib
 import id.walt.vclib.credentials.*
 import org.lighthousegames.logging.logging
 
-
 object Defaults {
 
     private val log = logging()
@@ -27,6 +26,7 @@ object Defaults {
     }
 
     fun loadVcLibDefaults() {
+
         // Register default types
         if (!defaults.isInitialized()) {
             log.info { "Registering default templates" }